is_multiple.py
This program requires the user to enter a number for m and a number for n. The function will determine of n is a multiple of m based on the user's input. If n is a multiple of m, the program returns True. This means that if n % m has a remainder of 0, n is in fact a multiple of m. If n is not a multiple of m, the program returns False.