repeat_sentence.py
This program will se a while loop to print a single sentence a hundred times. As integer x will be assigned 1 before the loop begins its execution. In the loop, each time the sentence is printed, x will be increased by 1. The loop continues to execute until x reach the value of 100.